feat(moderations): move /legacy to /moderations
