+++ src/bootstrap.ts
@@ -3,8 +3,11 @@
 import { csp_headers, type Csp_Context } from ":common/csp_headers";
 import env from ":common/env";
 // import { sentry, type Sentry_Context } from ":common/sentry";
+import { vip_list_guard } from ":auth/vip_list.guard";
+import { hyyyyyypertool_session } from ":common/session";
 import { moncomptepro_pg_database } from ":database:moncomptepro/middleware";
 import legacy from ":legacy/route";
+import { moderations_router } from ":moderations/route";
 import { sentry } from "@hono/sentry";
 import { Hono } from "hono";
 import { showRoutes } from "hono/dev";
@@ -47,6 +50,12 @@ const app = new Hono<Csp_Context>()
   .route("", auth_router)
   .route("", welcome_router)
   .use("*", moncomptepro_pg_database({ connectionString: env.DATABASE_URL }))
+  .use("*", hyyyyyypertool_session)
+  .use(
+    "/moderations/*",
+    vip_list_guard({ vip_list: env.ALLOWED_USERS.split(",") }),
+  )
+  .route("/moderations", moderations_router)
   .route("", legacy)
   .notFound(async ({ html, var: { nonce } }) => {
     return html(NotFound({ nonce }), 404);
