+++ src/legacy/routes/duplicate_warning/index.tsx
@@ -2,7 +2,7 @@
 
 import env from ":common/env";
 import { moncomptepro_pg_database } from ":database:moncomptepro/middleware";
-import { Duplicate_Warning } from ":legacy/moderations/Duplicate_Warning";
+import { Duplicate_Warning } from ":moderations/:id/Duplicate_Warning";
 import { zValidator } from "@hono/zod-validator";
 import { Hono } from "hono";
 import { z } from "zod";
