+++ src/moderations/:id/comment/Comment.tsx
@@ -1,9 +1,9 @@
 //
 
-import { UserInfo_Context } from ":auth/userinfo.context";
+import type { Session_Context } from ":common/session";
 import type { Moderation } from ":database:moncomptepro";
 import { input_group } from ":ui/form";
-import { useContext } from "hono/jsx";
+import { useRequestContext } from "hono/jsx-renderer";
 
 //
 
@@ -14,7 +14,11 @@ const MODERATION_COMMENT_ID = "moderation-comment";
 export function Comment({ moderation }: { moderation: Moderation }) {
   const { base, hint, input, label } = input_group({});
 
-  const userinfo = useContext(UserInfo_Context);
+  const {
+    var: { session },
+  } = useRequestContext<Session_Context>();
+  const userinfo = session.get("userinfo");
+  if (!userinfo) return <></>;
   const from = userinfo.email;
 
   return (
