+++ src/moderations/:id/page.tsx
@@ -10,8 +10,8 @@ import {
 } from ":database:moncomptepro";
 import { and, eq } from "drizzle-orm";
 import { createContext } from "hono/jsx";
-import { _02 } from "../02";
-import { _03 } from "../03";
+import { _02 } from "./02";
+import { _03 } from "./03";
 
 //
