+++ src/moderations/moderation_type_to_emoji.tsx
@@ -1,5 +1,9 @@
+//
+
 import type { MCP_Moderation } from ":moncomptepro";
 import { match } from "ts-pattern";
+
+//
 
 export function moderation_type_to_emoji(type: string) {
   return match(type as MCP_Moderation["type"])
